Refactored `rustcoalescence/algorithms/independent/*`
