+++ rustcoalescence/algorithms/independent/src/initialiser/genesis.rs
@@ -1,0 +1,90 @@
+use necsim_core::{
+    cogs::{EmigrationExit, MathsCore, PrimeableRng},
+    lineage::{GlobalLineageReference, Lineage},
+};
+
+use necsim_impls_no_std::cogs::{
+    active_lineage_sampler::independent::{
+        event_time_sampler::EventTimeSampler, IndependentActiveLineageSampler,
+    },
+    dispersal_sampler::in_memory::alias::InMemoryAliasDispersalSampler,
+    lineage_store::independent::IndependentLineageStore,
+    origin_sampler::TrustedOriginSampler,
+};
+
+use rustcoalescence_scenarios::Scenario;
+
+use super::IndependentLineageStoreSampleInitialiser;
+
+#[allow(clippy::module_name_repetitions)]
+pub struct GenesisInitialiser;
+
+#[allow(clippy::type_complexity)]
+impl<M: MathsCore, G: PrimeableRng<M>, O: Scenario<M, G>>
+    IndependentLineageStoreSampleInitialiser<M, G, O, !> for GenesisInitialiser
+{
+    type ActiveLineageSampler<
+        X: EmigrationExit<
+            M,
+            O::Habitat,
+            G,
+            GlobalLineageReference,
+            IndependentLineageStore<M, O::Habitat>,
+        >,
+        J: EventTimeSampler<M, O::Habitat, G, O::TurnoverRate>,
+    > = IndependentActiveLineageSampler<
+        M,
+        O::Habitat,
+        G,
+        X,
+        Self::DispersalSampler,
+        O::TurnoverRate,
+        O::SpeciationProbability,
+        J,
+    >;
+    type DispersalSampler = O::DispersalSampler<InMemoryAliasDispersalSampler<M, O::Habitat, G>>;
+
+    fn init<
+        'h,
+        T: TrustedOriginSampler<'h, M, Habitat = O::Habitat>,
+        J: EventTimeSampler<M, O::Habitat, G, O::TurnoverRate>,
+        X: EmigrationExit<
+            M,
+            O::Habitat,
+            G,
+            GlobalLineageReference,
+            IndependentLineageStore<M, O::Habitat>,
+        >,
+    >(
+        self,
+        origin_sampler: T,
+        dispersal_sampler: O::DispersalSampler<InMemoryAliasDispersalSampler<M, O::Habitat, G>>,
+        event_time_sampler: J,
+    ) -> Result<
+        (
+            IndependentLineageStore<M, O::Habitat>,
+            Self::DispersalSampler,
+            Self::ActiveLineageSampler<X, J>,
+            Vec<Lineage>,
+            Vec<Lineage>,
+        ),
+        !,
+    >
+    where
+        O::Habitat: 'h,
+    {
+        let (lineage_store, active_lineage_sampler, lineages) =
+            IndependentActiveLineageSampler::init_with_store_and_lineages(
+                origin_sampler,
+                event_time_sampler,
+            );
+
+        Ok((
+            lineage_store,
+            dispersal_sampler,
+            active_lineage_sampler,
+            lineages,
+            Vec::new(),
+        ))
+    }
+}
